Org: Only affect the address of the line that follows. (BC)
This makes more sense when labels scan forward and are directly associated with
the statement that follows it, otherwise a symbol would have the new address
even though it is on the line before. Also it allows use of $ to calculate the
new address from; previously it would not resolve.

Backwards incompatible change.