Line: Associate all preceding labels with a mnemonic.
And no longer emit lines without mnemonic from the parser.

With this we no longer need to put the mnemonic on the same line as the label
to let us access its inner scope, giving us a bit more formatting freedom.
Also improves the symbol file output.

The last line always has a terminator directive, so no labels are left over.
We had to adjust some tests for that though. And one additional test failed
because the label now referenced an instruction which could not be constructed.
It would already have failed previously had they been on the same line.