Parser: Change arguments scope to the parent.
It does not make sense for arguments to refer to their own line’s scope, when
the contents of said scope depend on it. This can create an infinite loop.
Additionally, the symbols generated in the line scope could shadow the symbols
that the arguments were referring to.

This caused issues such as exposed by the new tests.

There is a little wart in Identifier, because the context we pass in is the
line’s scope while now identifiers refer their parent. However I think the whole
context replacement is kinda iffy anyway, it replaces all contexts regardless of
what they were originally. Though it works because we only use it for arguments.