If: Improve expansion to match when other directives do it.
It used to register and expand later than usual in resolve / getBytes, because
we don’t want to evaluate the condition sooner in case $ is used in it.
However we now give the then and else sources their own scope, and register
symbols in the containing scope with the condition embedded in them.

This makes ifs work better inside macros.