ErrorLiteral: Introduce an error type.
This represents bottom from lambda calculus, a computation which never completes
successfully. It is a member of any type, and throws an exception on evaluation.

This allows some expressions to evaluate without throwing errors unnecessarily.
For example, scope lookups in a ternary which never evaluates to take the path
which causes the error.